Clean ScoredVector
