+++ nidx/nidx_vector/src/data_point_provider/reader.rs
@@ -19,9 +19,9 @@
 //
 
 use crate::data_point::OpenDataPoint;
-use crate::data_point::ScoredParagraph;
 use crate::data_point_provider::SearchRequest;
 use crate::data_point_provider::VectorConfig;
+use crate::data_store::ParagraphRef;
 use crate::request_types::VectorSearchRequest;
 use crate::utils;
 use crate::{VectorErr, VectorR};
@@ -34,6 +34,61 @@ use std::cmp::Ordering;
 use std::collections::{HashMap, HashSet};
 use std::time::Instant;
 use tracing::*;
+
+#[derive(Clone, Copy)]
+pub struct ScoredParagraph<'a> {
+    score: f32,
+    paragraph: ParagraphRef<'a>,
+}
+impl Eq for ScoredParagraph<'_> {}
+impl std::hash::Hash for ScoredParagraph<'_> {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        self.id().hash(state)
+    }
+
+    fn hash_slice<H: std::hash::Hasher>(data: &[Self], state: &mut H)
+    where
+        Self: Sized,
+    {
+        for piece in data {
+            piece.hash(state)
+        }
+    }
+}
+impl Ord for ScoredParagraph<'_> {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        self.paragraph.id().cmp(other.paragraph.id())
+    }
+}
+impl PartialOrd for ScoredParagraph<'_> {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+impl PartialEq for ScoredParagraph<'_> {
+    fn eq(&self, other: &Self) -> bool {
+        self.paragraph.id() == other.paragraph.id()
+    }
+}
+
+impl<'a> ScoredParagraph<'a> {
+    pub fn new(paragraph: ParagraphRef<'a>, score: f32) -> Self {
+        Self { paragraph, score }
+    }
+    pub fn score(&self) -> f32 {
+        self.score
+    }
+    pub fn id(&self) -> &str {
+        self.paragraph.id()
+    }
+    pub fn labels(&self) -> Vec<String> {
+        self.paragraph.labels()
+    }
+    pub fn metadata(&self) -> Option<&[u8]> {
+        let metadata = self.paragraph.metadata();
+        (!metadata.is_empty()).then_some(metadata)
+    }
+}
 
 // Fixed-sized sorted collection
 struct Fssc<'a> {
